pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = 'audio_spoof_detection_service'
version = "09.01.2026"
authors = [
     {name = "Sufiyarov Danil", email = "danil-sufiyarov@mail.ru"}
]
requires-python = ">= 3.12"
dependencies = [
    "fastapi[standard] == 0.128.0",
    "uvicorn == 0.40.0",
    "python-jose[cryptography]",
    "pyotp",
    "SQLAlchemy == 2.0.45",
    "torchaudio == 2.9.1",
    "torchcodec",
    "onnxruntime == 1.23.2",
    "dishka == 1.7.2",
    "python-dotenv == 1.2.1",
    "tqdm"
]
